debug mode

+++ lib/forklift/models/connection.rb
@@ -34,6 +34,7 @@ module Forklift
 
     def q(query, exit_on_exception=true)
       logger.debug "[ #{@name} ] #{query}"
+      return if Forklift::Debug.debug? == true
       begin
         response = connection.query(query);
         if hash_count(response) == 0
@@ -55,6 +56,7 @@ module Forklift
     end
 
     def connection_test
+      return if Forklift::Debug.debug? == true
       begin
         connection.query("select NOW()")
       rescue Error => e
@@ -93,6 +95,7 @@ module Forklift
         elsif local_connection.frequency_check(to, destination_table_name, forklift_data_table, frequency)
           start = Time.new.to_i
           logger.log " > importing remote table `#{from}`.`#{table}` to local `#{to}`.`#{destination_table_name}`"
+          return if Forklift::Debug.debug? == true
           local_connection.q("drop table if exists `#{to}`.`#{destination_table_name}`")
           create_table_command = connection.query("show create table #{table}").first["Create Table"]
           create_table_command.gsub!("CREATE TABLE `#{table}`", "CREATE TABLE `#{destination_table_name}`")
